% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodymap_df.R
\name{bodymap_df}
\alias{bodymap_df}
\title{Returns formatted data frame for SMMART Body Map plot}
\usage{
bodymap_df(
  df,
  col_name_date = NULL,
  col_name_lookup = NULL,
  col_name_desc = NULL,
  col_name_lat = NULL,
  col_name_type = NULL,
  remove_duplicate_sites = TRUE
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted on the body map.}

\item{col_name_date}{A string containing the column name of the column in `df` containing date
values. Optional, if df column is named "lesion_date".}

\item{col_name_lookup}{A string containing the column name of the column in `df` containing
values to join with the lesion_lookup dataset. Optional, if df column is named "lookup_name".}

\item{col_name_desc}{A string containing the column name of the column in `df` containing
lesion description values. Optional, if df column is named "site_description".}

\item{col_name_lat}{A string containing the column name of the column in `df` containing
lesion laterality values. Optional, if df column is named "laterality".}

\item{col_name_type}{A string containing the column name of the column in `df` containing
lesion type values (Primary, Metastatic). Optional, if df column is named "lesion_type".}

\item{remove_duplicate_sites}{Logical value indicating whether duplicate lesion sites should
be removed from data frame. Default is TRUE.}
}
\value{
Data Frame
}
\description{
Returns formatted data frame for SMMART Body Map plot
}
\examples{
bodymap_df(example_patient_dataset)
}
