% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodymap_plot.R
\name{bodymap_plot}
\alias{bodymap_plot}
\title{bodymap_plot}
\usage{
bodymap_plot(df, sex)
}
\arguments{
\item{df}{Data frame formatted using bodymap_data function.}

\item{sex}{String value indicating sex of image to plot against. Choices are "female" or "male"}
}
\value{
ggplot of lesion locations
}
\description{
bodymap_plot
}
\examples{
df <- bodymap_df(example_patient_dataset)
bodymap_plot(df, 'female')
}
