% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_patient_data}
\alias{example_patient_data}
\title{Example patient data}
\format{
## `example_patient_data`
A data frame with 14 rows and 5 columns:
\describe{
  \item{lesion_date}{Date of lesion diagnosis}
  \item{lookup_name}{Value to match against `lesion_lookup` site list.
  May be the same value as `site_description`.}
  \item{site_description}{Descriptive text for lesion site used in plot labels}
  \item{laterality}{Side of body on which lesion occurred (Bilateral, Left, Midline, Right,
  Not a paired site, Unknown)}
  \item{lesion_type}{Identifies lesion as primary or metastatic (Primary, Metastatic)}
}
}
\description{
A data frame of patient lesion data, formatted for input to `bodymap_df`.
}
