% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesion_lookup}
\alias{lesion_lookup}
\title{Lookup list for lesion plot locations}
\format{
## `lesion_lookup`
A data frame with 1,361 rows and 7 columns:
\describe{
  \item{Type}{Identifies lesion as primary or metastatic (Primary, Metastatic)}
  \item{Site}{Value to match against `lookup_name` input column. Primary sites use ICD-O-3
  morphology terms. Metastatic sites use custom site terms.}
  \item{ICD Code}{ICD-O-3 morphology code. Not currently used for matching; provided
  as reference.}
  \item{y}{Y plot location value.}
  \item{x, x_left,x_right}{X plot location. Lesions with lateralities of Midline,
  Not a paired site, and Unknown use x. Otherwise, left or right location used.}
}
}
\source{
Lesion site development source: https://github.com/TheMillerLab/BodyMapR Plot locations
reviewed for alignment with modified body map. Manual adjustments performed as required.
}
\usage{
lesion_lookup
}
\description{
#' @description
}
\details{
A list of lesion sites for both primary and metastatic lesion types. Used for matching against
 input data to determine plot location for input lesions.
}
\keyword{datasets}
